Move findKeyringForAddress to metamask duck
